+++ src/command/question/index.ts
@@ -2,3 +2,4 @@ export * from './Boilerplate';
 export * from './Diff';
 export * from './Preset';
 export * from './Project';
+export * from './Channel';